requirements.txt
# Core
Django>=6.0,<7.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
psutil>=5.9

# Development
django-debug-toolbar>=4.0
